moved displaycontroller
